Fixed if-then-else

+++ hive/lit/misc/api.h
@@ -32,5 +32,5 @@ double clampProb(double p);
 void nsToTs(uint64_t ns, struct timespec * pTs);
 void sleepNs(uint64_t ns);
 void sleepMs(uint64_t ms);
-bool rollCumGauss(double x, double mu, double sigma);
+bool rollCumGauss(float x, float mu, float sigma);
