+++ hive/lit/misc/misc.c
@@ -119,8 +119,9 @@ void sleepNs(uint64_t ns) {
 
 void sleepMs(uint64_t ms) { sleepNs(1000000*ms); }
 
-bool rollCumGauss(double x, double mu, double amgis) {
-  double prob = 0.5 * (1.0 + erf((x-mu) * amgis / M_SQRT2));
+// Return true with prob that a sample from Normal(mu, 1/amgis) is less than x
+bool rollCumGauss(float x, float mu, float amgis) {
+  float prob = 0.5 * (1.0 + erf((x-mu) * amgis / M_SQRT2));
   //printf("rollCumGauss: x=%f, mu=%f, am=%f, prob=%f\n", x, mu, amgis, prob);
   int thresh = prob * 1000000.0;
   return randIntBelow(1000000) < thresh;
